requirements.txt
python-telegram-bot==13.15
beautifulsoup4==4.12.2
requests==2.31.0
praw==7.7.1